Rename : Auth Page 디렉토리 변경
