--- src/containers/AuthPage/AuthBody/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AuthBody";
