--- src/containers/AuthPage/index.js
@@ -1,2 +1,0 @@
-export { default as AuthHeader } from "./AuthHeader";
-export { default as AuthBody } from "./AuthBody";
