+++ src/containers/index.js
@@ -4,4 +4,3 @@ export * from "./CommunityPage";
 export * from "./JobPage";
 export * from "./StudyPage";
 export * from "./MyPage";
-export * from "./AuthPage";
